# refactor: 10x faster RPC splitting