--- a/gossipsub_test.go
+++ b/gossipsub_test.go
@@ -8,6 +8,7 @@ import (
 	"fmt"
 	"io"
 	mrand "math/rand"
+	mrand2 "math/rand/v2"
 	"sort"
 	"strings"
 	"sync"
@@ -3673,5 +3674,30 @@ func TestPublishDuplicateMessage(t *testing.T) {
 	err = topic.Publish(ctx, []byte("hello"))
 	if err != nil {
 		t.Fatal("Duplicate message should not return an error")
+	}
+}
+
+func genNRpcs(tb testing.TB, n int, maxSize int) []*RPC {
+	r := mrand2.NewChaCha8([32]byte{})
+	rpcs := make([]*RPC, n)
+	for i := range rpcs {
+		var data [64]byte
+		_, err := r.Read(data[:])
+		if err != nil {
+			tb.Fatal(err)
+		}
+		rpcs[i] = generateRPC(data[:], maxSize)
+	}
+	return rpcs
+}
+
+func BenchmarkSplitRPC(b *testing.B) {
+	maxSize := 2048
+	rpcs := genNRpcs(b, 100, maxSize)
+
+	b.ResetTimer()
+	for i := 0; i < b.N; i++ {
+		rpc := rpcs[i%len(rpcs)]
+		appendOrMergeRPC(nil, maxSize, *rpc)
 	}
 }