Builds on #582. 10x faster than current master. 0 allocs.

The basic logic is the same as the old version, except we return an `iter.Seq[RPC]` and yield `RPC` types instead of a slice of `*RPC`. This lets us avoid allocations for heap pointers.

Please review @algorandskiy, and let me know if this improves your use case.