@MarcoPolo 2025-05-24
benchstat:

```
goos: darwin
goarch: arm64
pkg: github.com/libp2p/go-libp2p-pubsub
cpu: Apple M1 Max
                                             │   old.txt    │               new.txt               │
                                             │    sec/op    │   sec/op     vs base                │
SplitRPCLargeMessages/Many_large_messages-10   51.751µ ± 0%   4.327µ ± 1%  -91.64% (p=0.000 n=10)
SplitRPCLargeMessages/2_large_messages-10      320.70n ± 2%   34.63n ± 0%  -89.20% (p=0.000 n=10)
geomean                                         4.074µ        387.0n       -90.50%

                                             │   old.txt    │                 new.txt                  │
                                             │     B/op     │    B/op      vs base                     │
SplitRPCLargeMessages/Many_large_messages-10   18.65Ki ± 0%   0.00Ki ± 0%  -100.00% (p=0.000 n=10)
SplitRPCLargeMessages/2_large_messages-10        392.0 ± 0%      0.0 ± 0%  -100.00% (p=0.000 n=10)
geomean                                        2.672Ki                     ?                       ¹ ²
¹ summaries must be >0 to compute geomean
² ratios must be >0 to compute geomean

                                             │  old.txt   │                 new.txt                 │
                                             │ allocs/op  │ allocs/op   vs base                     │
SplitRPCLargeMessages/Many_large_messages-10   217.0 ± 0%     0.0 ± 0%  -100.00% (p=0.000 n=10)
SplitRPCLargeMessages/2_large_messages-10      8.000 ± 0%   0.000 ± 0%  -100.00% (p=0.000 n=10)
geomean                                        41.67                    ?                       ¹ ²
¹ summaries must be >0 to compute geomean
² ratios must be >0 to compute geomean
```